+++ v7/scripts/update-emulators.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 
-yarn add emulators emulators-ui
+yarn add emulators emulators-ui js-dos@beta
 rm -rf static/releases/latest
 mkdir static/releases/latest
 mkdir static/releases/latest/emulators
 mkdir static/releases/latest/emulators-ui
+mkdir static/releases/latest/js-dos
+
 cp -r node_modules/emulators/dist/* static/releases/latest/emulators
 cp -r node_modules/emulators-ui/dist/* static/releases/latest/emulators-ui
+cp -r node_modules/js-dos/dist/* static/releases/latest/js-dos