+++ v7/static/releases/latest/emulators-ui/types/controls/nipple-arrows.d.ts
@@ -1,0 +1,3 @@
+import { CommandInterface } from "emulators";
+import { Layers } from "../dom/layers";
+export declare function nippleArrows(layers: Layers, ci: CommandInterface): void;