+++ v7/static/releases/latest/emulators-ui/types/emulators-ui.d.ts
@@ -2,8 +2,9 @@ import { layers } from "./dom/layers";
 import { resolveBundle } from "./network/xhr";
 import { webGl } from "./graphics/webgl";
 import { keyboard } from "./controls/keyboard";
+import { nippleArrows } from "./controls/nipple-arrows";
 import { audioNode } from "./sound/audio-node";
-export declare const EmulatorsUi: {
+export declare class EmulatorsUi {
     dom: {
         layers: typeof layers;
     };
@@ -18,5 +19,6 @@ export declare const EmulatorsUi: {
     };
     controls: {
         keyboard: typeof keyboard;
+        nippleArrows: typeof nippleArrows;
     };
-};
+}